backend/requirements.txt
fastapi
uvicorn[standard]
supabase
plaid-python
anthropic
openai
python-jose[cryptography]
pydantic-settings
httpx
python-dotenv
pytest
